Add a '--mop-up' command line option
At the end of the shutdown, there are donors with unused donations
and recipients not full, because we try not to assign donations to a
donor if we can't do it all in one wave.  The '--mop-up' option pauses
this behavior in order to get everyone as many donations as possible
in the current wave.